修改类型

+++ models/badDebtRecord.go
@@ -7,9 +7,9 @@ const TableNameBadDebtRecord = "bad_debt_record"
 // BadDebtRecord 坏账记录表
 type BadDebtRecord struct {
 	BadDebtId        int64           `json:"bad_debt_id" gorm:"bad_debt_id,primaryKey"`    // 主键ID
-	MerchantId       int             `json:"merchant_id" gorm:"merchant_id"`               // 商户ID
+	MerchantId       int64           `json:"merchant_id" gorm:"merchant_id"`               // 商户ID
 	UserId           int64           `json:"user_id" gorm:"user_id"`                       // 用户ID
-	PlatformId       int             `json:"platform_id" gorm:"platform_id"`               // 平台ID
+	PlatformId       int64           `json:"platform_id" gorm:"platform_id"`               // 平台ID
 	DebtAmount       decimal.Decimal `json:"debt_amount" gorm:"debt_amount"`               // 坏账金额（应扣除但无法扣除的金额）
 	UserBalance      decimal.Decimal `json:"user_balance" gorm:"user_balance"`             // 用户当前余额
 	DeficitAmount    decimal.Decimal `json:"deficit_amount" gorm:"deficit_amount"`         // 缺口金额（负余额的绝对值）
